+++ src/components/button.js
@@ -1,8 +1,8 @@
 import {AbstractComponent} from './abstract.js';
-class Button extends AbstractComponent  {
+class Button extends AbstractComponent {
   constructor() {
     super();
-    }
+  }
   getTemplate() {
     return `<button class="films-list__show-more">Show more</button>`;
   }
