+++ src/components/film-container.js
@@ -1,21 +1,9 @@
-import {createElement} from "../utils.js";
+import {AbstractComponent} from './abstract.js';
 
-class FilmContainer {
+class FilmContainer extends AbstractComponent {
   constructor() {
+    super();
   }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
-
   getTemplate() {
     return `<section class="films">
         <section class="films-list">
