+++ src/components/top-rated.js
@@ -1,22 +1,8 @@
-import {createElement} from "../utils.js";
-
-class TopRated {
+import {AbstractComponent} from './abstract.js';
+class TopRated extends AbstractComponent {
   constructor() {
-    this._element = null;
+    super();
   }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
-
   getTemplate() {
     return `<section class="films-list--extra">
           <h2 class="films-list__title">Top rated</h2>
